Added check to ensure we grabbed the correct table
During configuration If a user's ratio is dangerously low the page displays an extra table with a warning banner.  This extra table throws off the jQuery selection.  Since the table does not have any identifiers to write a better selector, I just ensure that the amount of rows returned isn't 4, if it is we select the next table.